feat: rename to styled components files
